Улучшение отладки [+] Разделение шагов вперёд: Шаг с заходом и Шаг с обходом [+] Добавление кнопки на гланый экран и в клавиши быстрого доступа

+++ Emulator_Intel8080/Source/Core/Emulator/EmulationThread/EmulationThread.cpp
@@ -366,7 +366,7 @@ void EmulationThread::DrawMainMenu() {
 
 	if (glfwGetTime() - LastTimeOneSecond > .25f) {
 
-		LastTimeOneSecond = glfwGetTime();
+		LastTimeOneSecond = (float)glfwGetTime();
 
 		if (LastCountTicksInSeconds_temp > processor->GetCountTicks()) {
 			LastCountTicksInSeconds = 0;
